requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiogram==3.4.1
Jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.0 